test: refactor duplicate inplace pod resize test utilities
